feat: Add title prop to ExportButton
This commit introduces a `title` prop to the ExportButton
component. This allows users to specify a custom title for
the exported file, overriding the default filename or the
fallback "export" value. This improves the user experience
by providing more control over the naming of exported files.
A default style of `w-full sm:w-auto` was added to the button.